bugado

+++ src/main/java/com/backend/techstock/dto/SalesDto.java
@@ -1,9 +1,10 @@
 package com.backend.techstock.dto;
 
 import java.time.LocalDateTime;
+import java.math.BigDecimal;
 
 public record SalesDto( 
               String saleName, 
               String saleDescription,
-              double saleDiscount,
+              BigDecimal saleDiscount,
               LocalDateTime saleDate) {}
